+++ packages/iov-tendermint-rpc/src/responses.ts
@@ -119,7 +119,7 @@ export interface TxResponse {
   readonly hash: TxHash;
   readonly height: number;
   readonly index: number;
-  readonly txResult: TxData;
+  readonly result: TxData;
   readonly proof?: TxProof;
 }